refactor callbacks
